Researcher работает, Copier.swapAll работает

+++ src/main/java/com/group/original/panopticon/file/Copier.java
@@ -1,5 +1,6 @@
 package com.group.original.panopticon.file;
 
+import com.group.original.panopticon.StandardPaths;
 import com.group.original.panopticon.file.differences.Differences;
 import com.group.original.panopticon.file.system.FileStamp;
 
@@ -51,13 +52,15 @@ public class Copier {
 
         String fileName = fileStamp.getRelativePath().getFileName().toString();
         String oldFileName = fileStamp.getFormattedLastModifiedTime().concat("-").concat(fileName);
-        Path oldVersion = Path.of(oldFileName);
+        Path oldVersion = StandardPaths.getOldVersionsPath().resolve(oldFileName);
 
         try {
             Files.deleteIfExists(oldVersion);
-            Files.copy(targetPath, oldVersion);
+            if (Files.exists(targetPath)) {
+                Files.copy(targetPath, oldVersion);
+            }
 
-            Files.delete(targetPath);
+            Files.deleteIfExists(targetPath);
             Files.copy(sourcePath, targetPath);
         } catch (IOException ioException) {
             ioException.printStackTrace();
